Conflicts:
	src/open/test-core/test_core
	src/restricted/applications/calibrator/calibratorMainWindow.cpp
	src/restricted/applications/calibrator/calibratorMainWindow.h
	src/restricted/applications/calibrator/inputDirectory.cpp
	src/restricted/applications/calibrator/inputDirectory.h

+++ test/serialize1/main_serialize1.cpp
@@ -169,7 +169,7 @@ void testJSON2()
     }
 
     ASSERT_TRUE(testObject.testField.size() == testObject1.testField.size(), "Wrong loaded array size");
-    for (int i = 0; i < testObject.testField.size(); i++) {
+    for (size_t i = 0; i < testObject.testField.size(); i++) {
         ASSERT_TRUE_P(testObject1.testField[i] == testObject.testField[i], ("Error at pos %d", i));
         cout << testObject1.testField[i] << " == " << testObject.testField[i] << std::endl;
     }
